fix: fix bugs

+++ bekend/src/main/java/com/application/bekend/service/CompanyService.java
@@ -52,7 +52,7 @@ public class CompanyService {
 		return this.userCategoryService.addCategory(dto);
 	}
 
-	public boolean deleteCategory(int id) {
+	public boolean deleteCategory(Long id) {
 		return this.userCategoryService.delete(id);
 	}
